+++ versioned/storage/dynamodb/src/main/java/org/projectnessie/versioned/storage/dynamodb/DynamoDBConstants.java
@@ -29,6 +29,7 @@ final class DynamoDBConstants {
   static final String COL_REFERENCES_DELETED = "d";
   static final String COL_REFERENCES_CREATED_AT = "c";
   static final String COL_REFERENCES_EXTENDED_INFO = "e";
+  static final String COL_REFERENCES_PREVIOUS = "h";
 
   static final String COL_REFERENCES_CONDITION_COMMON =
       "(" + COL_REFERENCES_DELETED + " = :deleted) AND (" + COL_REFERENCES_POINTER + " = :pointer)";